Adopt to moved packages in Elasticsearch
